Added SourceItem to Item class to contain a direct reference to the Plugin Item associated with each Model Item.  The FQN to the source Item is stored in the configuration file and is resolved at runtime.
This change allows for the new ReadFromSource() method in the Item class; allowing external code to force a model item to resolve a new value directly from whatever plugin (or other model item) it is reading from.

When calculated tags are introduced this will take some thought to work around; perhaps the SourceAddress/Item fields will be null and the calculation will Write() data into the model item instead of it being bound to a calculation somehow.